jsonファイルを扱う関数を外部に切り出すため
