+++ 04.class/memo.js
@@ -1,4 +1,9 @@
-import fs from "fs";
+import {
+  addDataToJsonFile,
+  deleteDataToJsonFile,
+  readJsonFile,
+} from "./jsonFileHandleFunctions.js";
+
 import { v4 as uuidv4 } from "uuid";
 
 export default class Memo {
@@ -41,31 +46,4 @@ export default class Memo {
   destroy() {
     deleteDataToJsonFile(Memo.storagePath, this.id);
   }
-}
-
-function addDataToJsonFile(path, newData) {
-  const jsonData = readJsonFile(path);
-
-  jsonData.push(newData);
-
-  fs.writeFileSync(path, JSON.stringify(jsonData), "utf8");
-}
-
-function deleteDataToJsonFile(path, id) {
-  const jsonData = readJsonFile(path);
-
-  const deletedJsonData = jsonData.filter((data) => {
-    return data.id !== id;
-  });
-
-  fs.writeFileSync(path, JSON.stringify(deletedJsonData), "utf8");
-}
-
-function readJsonFile(path) {
-  if (!fs.existsSync(path)) {
-    return [];
-  }
-
-  const fileData = fs.readFileSync(path, "utf8");
-  return JSON.parse(fileData);
 }
